TwitchHandler: Added verification ID checks and handles uninitalized econData for the user
